refactor: Moved all auth related routes in a single auth blue print
