+++ admin/backend/app.py
@@ -47,6 +47,7 @@ def create_app(bench_root: Path) -> Flask:
 def register_blueprints(app: Flask) -> None:
     from admin.backend.api.v1.apps import apps_bp, marketplace_bp
     from admin.backend.api.v1.benches import bench_readiness_bp, benches_bp
+    from admin.backend.api.v1.auth import auth_bp
     from admin.backend.api.v1.core import core_bp
     from admin.backend.api.v1.databases import database_bp
     from admin.backend.api.v1.editor import editor_bp
@@ -54,7 +55,6 @@ def register_blueprints(app: Flask) -> None:
     from admin.backend.api.v1.logs import logs_bp
     from admin.backend.api.v1.migrations import migrations_bp
     from admin.backend.api.v1.processes import processes_bp
-    from admin.backend.api.v1.sessions import sessions_bp
     from admin.backend.api.v1.settings import audit_bp, network_bp, settings_bp
     from admin.backend.api.v1.setup import setup_bp
     from admin.backend.api.v1.sites import sites_bp
@@ -76,7 +76,7 @@ def register_blueprints(app: Flask) -> None:
     app.register_blueprint(tasks_bp, url_prefix=f"{API_V1_PREFIX}/tasks")
     app.register_blueprint(task_worker_bp, url_prefix=API_V1_PREFIX)
     app.register_blueprint(settings_bp, url_prefix=f"{API_V1_PREFIX}/settings")
-    app.register_blueprint(sessions_bp, url_prefix=f"{API_V1_PREFIX}/sessions")
+    app.register_blueprint(auth_bp, url_prefix=f"{API_V1_PREFIX}/auth")
     app.register_blueprint(audit_bp, url_prefix=API_V1_PREFIX)
     app.register_blueprint(network_bp, url_prefix=API_V1_PREFIX)
     app.register_blueprint(updates_bp, url_prefix=API_V1_PREFIX)
