+++ admin/frontend/dashboard/src/api/session.js
@@ -1,7 +1,7 @@
 import { request, unwrap } from './client'
 
 export const sessionApi = {
-  list: () => unwrap(request.get('sessions').json()),
-  revoke: (jti) => request.post(`sessions/revoke/${jti}`),
-  revokeAll: () => unwrap(request.post('sessions/revoke/all').json()),
+  list: () => unwrap(request.get('auth/sessions').json()),
+  revoke: (jti) => request.post(`auth/sessions/revoke/${jti}`),
+  revokeAll: () => unwrap(request.post('auth/sessions/revoke/all').json()),
 }
