+++ admin/frontend/dashboard/src/api/settings.js
@@ -3,7 +3,7 @@ import { request, unwrap } from './client'
 export const settingsApi = {
   get: () => unwrap(request.get('settings').json()),
   update: (data) => unwrap(request.patch('settings', { json: data }).json()),
-  changeAdminPassword: (data) => unwrap(request.post('settings/admin-password', { json: data }).json()),
+  changeAdminPassword: (data) => unwrap(request.post('auth/password', { json: data }).json()),
   myIp: () => request.get('network/client').json(),
   llmModels: (provider) =>
     request.get('settings/llm/models', { searchParams: { provider } }).json(),
