+++ tests/admin/backend/api/test_admin_routes.py
@@ -107,13 +107,12 @@ def test_admin_route_inventory_matches_baseline(tmp_path: Path) -> None:
         "network": 1,
         "updates": 1,
         "runtime": 4,
-        "settings": 8,
+        "auth": 11,
+        "settings": 3,
         "setup": 6,
         "sites": 33,
         "ssh-keys": 3,
         "metrics": 1,
-        "session": 3,
-        "sessions": 3,
         "system": 1,
         "task-worker": 3,
         "tasks": 8,
@@ -208,11 +207,11 @@ def test_admin_route_inventory_matches_baseline(tmp_path: Path) -> None:
         ("GET", "/api/v1/settings"),
         ("PATCH", "/api/v1/settings"),
         ("GET", "/api/v1/settings/llm/models"),
-        ("POST", "/api/v1/settings/admin-password"),
-        ("GET", "/api/v1/settings/two-factor"),
-        ("POST", "/api/v1/settings/two-factor/enrollment"),
-        ("POST", "/api/v1/settings/two-factor/<credential_id>"),
-        ("DELETE", "/api/v1/settings/two-factor/<credential_id>"),
+        ("POST", "/api/v1/auth/password"),
+        ("GET", "/api/v1/auth/two-factor"),
+        ("POST", "/api/v1/auth/two-factor/enrollment"),
+        ("POST", "/api/v1/auth/two-factor/<credential_id>"),
+        ("DELETE", "/api/v1/auth/two-factor/<credential_id>"),
         ("GET", "/api/v1/audit-events"),
         ("GET", "/api/v1/network/client"),
         ("GET", "/api/v1/ssh-keys"),
