+++ tests/admin/backend/test_jwks.py
@@ -201,9 +201,9 @@ def test_jwks_ec_bearer_token_authenticates(tmp_path: Path) -> None:
 
 def test_jwks_sid_login_with_jti_sets_cookie(tmp_path: Path) -> None:
     client = _client(tmp_path)
-    resp = client.post("/api/v1/session", json={"sid": _mint(jti="login-1")})
+    resp = client.post("/api/v1/auth/session", json={"sid": _mint(jti="login-1")})
     assert resp.status_code == 201
-    assert resp.headers["Location"] == "/api/v1/session"
+    assert resp.headers["Location"] == "/api/v1/auth/session"
     assert client.get("/api/v1/benches").status_code != 401
 
 
@@ -211,14 +211,14 @@ def test_jwks_sid_login_requires_jti(tmp_path: Path) -> None:
     # A token without a jti must not be exchangeable for a session (else it is
     # replayable until expiry).
     client = _client(tmp_path)
-    assert client.post("/api/v1/session", json={"sid": _mint()}).status_code == 401
+    assert client.post("/api/v1/auth/session", json={"sid": _mint()}).status_code == 401
 
 
 def test_jwks_sid_login_is_single_use(tmp_path: Path) -> None:
     client = _client(tmp_path)
     sid = _mint(jti="login-2")
-    assert client.post("/api/v1/session", json={"sid": sid}).status_code == 201
-    assert client.post("/api/v1/session", json={"sid": sid}).status_code == 401
+    assert client.post("/api/v1/auth/session", json={"sid": sid}).status_code == 201
+    assert client.post("/api/v1/auth/session", json={"sid": sid}).status_code == 401
 
 
 def test_jwks_sid_login_without_exp_does_not_crash(tmp_path: Path) -> None:
@@ -229,13 +229,13 @@ def test_jwks_sid_login_without_exp_does_not_crash(tmp_path: Path) -> None:
         algorithm="RS256",
         headers={"kid": "rsa-key"},
     )
-    assert client.post("/api/v1/session", json={"sid": forever}).status_code == 401
+    assert client.post("/api/v1/auth/session", json={"sid": forever}).status_code == 401
 
 
 def test_jwks_site_scoped_token_cannot_bootstrap_session(tmp_path: Path) -> None:
     # A site-scoped token (even with a jti) must not escalate to a bench session.
     client = _client(tmp_path)
-    resp = client.post("/api/v1/session", json={"sid": _mint(jti="login-3", scope="site", site="a.com")})
+    resp = client.post("/api/v1/auth/session", json={"sid": _mint(jti="login-3", scope="site", site="a.com")})
     assert resp.status_code == 401
