+++ tests/integration/test_production_ssl.py
@@ -352,7 +352,7 @@ class TestProductionSSL:
 
     def test_admin_login_works(self, production: Path) -> None:
         status, body = _request(
-            ADMIN_DOMAIN, "/api/v1/session", method="POST", json_body={"password": ADMIN_PASSWORD}
+            ADMIN_DOMAIN, "/api/v1/auth/session", method="POST", json_body={"password": ADMIN_PASSWORD}
         )
         assert status == "201", f"/api/v1/session returned {status}: {body!r}"
         assert json.loads(body).get("authenticated") is True, f"login failed: {body!r}"
